Reorganize project structure
The `streaming` module was splat into three modules:
- consumers
- graph_builders
- processors
