+++ plugin/arcane-stream-sqlserver-change-tracking/integration-tests.env
@@ -1,4 +1,4 @@
-STREAMCONTEXT__BACKFILL=false
+STREAMCONTEXT__BACKFILL=true
 STREAMCONTEXT__SPEC='{ "database": "IntegrationTests", "schema": "dbo", "table": "TestTable", "changeCaptureIntervalSeconds": 1, "commandTimeout": 3600, "groupingIntervalSeconds": 3, "groupsPerFile": 1, "lookBackInterval": 3600, "rowsPerGroup": 10000, "sinkLocation": "integration_tests_target_table", "catalogSettings": { "namespace": "test", "warehouse": "polaris", "catalogUri": "http://localhost:8181/api/catalog" } }'
 STREAMCONTEXT__STREAM_ID=test
 STREAMCONTEXT__STREAM_KIND=SqlServerChangeTracking
